some refactoring

+++ server/views/detections.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, get_object_or_404
 from django.core.paginator import Paginator
 from django.http import JsonResponse
 from django.contrib.auth.decorators import login_required
+from django.db.models import F
 from rest_framework.parsers import JSONParser
 from ..models import Detection, Person
 from ..utils.storage import delete_file
@@ -9,7 +10,9 @@ from ..utils.storage import delete_file
 
 def _get_detections(request, verified, render_page, person_id):
     detections = Detection.objects.filter(verified=verified).order_by('person__name')
-    persons = Person.objects.all()
+    names = detections.values_list('person__name', flat=True).distinct()
+    persons = Person.objects.filter(name__in=names)
+
     ctx = { 'elements': detections, 'persons': persons }
 
     if person_id:
@@ -17,7 +20,7 @@ def _get_detections(request, verified, render_page, person_id):
         ctx['elements'] = detections
         ctx['chosen'] = person_id
 
-    paginator = Paginator(detections, 1)
+    paginator = Paginator(detections, 4)
     page = request.GET.get('page')
 
     if not page:
@@ -34,9 +37,16 @@ def _get_detections(request, verified, render_page, person_id):
 
 def _delete_detection(request):
     data = JSONParser().parse(request)
-    detection_id = data['id']
+    detection_id = data.get('id')
 
-    detection = Detection.objects.get(id=detection_id)
+    if not detection_id:
+        return JsonResponse({ 'error': True })
+
+    try:
+        detection = Detection.objects.get(id=detection_id)
+    except Detection.DoesNotExist:
+        return JsonResponse({ 'error': True })
+    
     image = detection.image
     detection.delete()
     delete_file(image.image_file.path)
@@ -58,8 +68,16 @@ def verified(request, person_id=None):
     if request.method == 'GET':
         return _get_detections(request, True, 'verified.html', person_id)
     elif request.method == 'POST':
-        id = JSONParser().parse(request)['id']
-        detection = Detection.objects.get(id=id)
+        id = JSONParser().parse(request).get('id')
+        
+        if not id:
+            return JsonResponse({ 'error': True })
+
+        try:
+            detection = Detection.objects.get(id=id)
+        except Detection.DoesNotExist:
+            return JsonResponse({ 'error': True })
+            
         detection.verified = True
         detection.save()
